clippy: clean -D warnings on workspace for phase-a3
Fixes:
- iss.rs: doc_lazy_continuation - blank doc-comment line ends the
  bullet list before the trailing paragraph
- engine.rs reap_time_wait: unnecessary_map_or -> is_some_and
- resd-net/src/lib.rs: unnecessary_cast on ConnHandle (*conn as u32)
- tcp_basic_tap.rs: field_reassign_with_default -> struct literal
  with ..Default::default()
- tcp_basic_tap.rs: never_loop -> .incoming().next() with if let

Verified: cargo clippy --workspace --all-targets -- -D warnings
exits 0; cargo test -p resd-net-core -> 108 unit + 3 integration
tests pass.

+++ crates/resd-net/src/lib.rs
@@ -136,7 +136,7 @@ pub unsafe extern "C" fn resd_net_poll(
                 // last_read_buf at the event's byte_offset (Task 19 fix
                 // for multi-segment polls).
                 let ft = engine.flow_table();
-                let (data_ptr, data_len) = match ft.get(*conn as u32) {
+                let (data_ptr, data_len) = match ft.get(*conn) {
                     Some(c) => {
                         let off = *byte_offset as usize;
                         let ptr = unsafe { c.recv.last_read_buf.as_ptr().add(off) };
